Rename `direct::with_async` to `direct::future` and `async_sink` to `sinks`
These read much more reasonable.
